More platform independency introduced.

+++ src/core/fstream.cc
@@ -406,7 +406,8 @@ namespace nil {
                     if ((buf.size() & (_file.disk_write_dma_alignment() - 1)) != 0) {
                         // If buf size isn't aligned, copy its content into a new aligned buf.
                         // This should only happen when the user calls output_stream::flush().
-                        auto tmp = allocate_buffer(align_up(buf.size(), _file.disk_write_dma_alignment()));
+                        auto tmp = allocate_buffer(
+                            align_up(buf.size(), static_cast<size_t>(_file.disk_write_dma_alignment())));
                         ::memcpy(tmp.get_write(), buf.get(), buf.size());
                         ::memset(tmp.get_write() + buf.size(), 0, tmp.size() - buf.size());
                         buf = std::move(tmp);
